tests: Clean up tests folder
