--- tests/asr/test_ftspeech.py
@@ -1,1 +1,0 @@
-"""Unit tests for the `asr.ftspeech` module."""
