--- tests/asr/test_ngram_model.py
@@ -1,1 +1,0 @@
-"""Unit tests for the `asr.ngram_model` module."""
