--- tests/asr/test_utils.py
@@ -1,1 +1,0 @@
-"""Unit tests for the `asr.utils` module."""
