--- tests/asr/wav2vec2/test_compute_metrics.py
@@ -1,1 +1,0 @@
-"""Unit tests for the `asr.wav2vec2.compute_metrics` module."""
