--- tests/asr/wav2vec2/test_data_collator.py
@@ -1,1 +1,0 @@
-"""Unit tests for the `asr.wav2vec2.data_collator` module."""
