--- tests/asr/wav2vec2/test_evaluate.py
@@ -1,1 +1,0 @@
-"""Unit tests for the `asr.wav2vec2.evaluate` module."""
