+++ tests/compute_metrics.py
@@ -1,0 +1,1 @@
+"""Unit tests for the `compute_metrics` module."""
