+++ tests/model_setup.py
@@ -1,0 +1,1 @@
+"""Unit tests for the `model_setup` module."""
